Running tests
Just testing the two audio-to-midi methods
